server: move all copy functions into same file
